+++ pyonepassword/pyonepassword.py
@@ -1,6 +1,7 @@
+import fnmatch
 import logging
 from os import environ as env
-from typing import Optional, Type, Union
+from typing import List, Optional, Type, Union
 
 from ._py_op_commands import (
     EXISTING_AUTH_IGNORE,
@@ -36,7 +37,9 @@ from .py_op_exceptions import (
     OPInvalidDocumentException,
     OPInvalidItemException,
     OPItemDeleteException,
+    OPItemDeleteMultipleException,
     OPItemGetException,
+    OPItemListException,
     OPSignoutException
 )
 from .version import PyOPAboutMixin
@@ -539,7 +542,7 @@ class OP(_OPCommandInterface, PyOPAboutMixin):
 
         return document_id
 
-    def item_list(self, categories=[], include_archive=False, tags=[], vault=None, generic_okay=True) -> OPItemList:
+    def item_list(self, categories=[], include_archive=False, tags=[], title_glob=None, vault=None, generic_okay=True) -> OPItemList:
         """
         Return a list of items in an account.
 
@@ -551,6 +554,10 @@ class OP(_OPCommandInterface, PyOPAboutMixin):
             Include items in the Archive in the list
         tags: List[str], optional
             A list of tags to restrict list to
+        title_glob: bool, optional
+            a shell-style glob pattern to match against item titles. If provided,
+            resulting list will include only matching items
+            by default None
         vault: str, optional
             The name or ID of a vault to override the object's default vault
         generic_okay: bool, optional
@@ -574,6 +581,13 @@ class OP(_OPCommandInterface, PyOPAboutMixin):
         item_list_json = self._item_list(
             categories, include_archive, tags, vault)
         item_list = OPItemList(item_list_json, generic_okay=generic_okay)
+
+        if title_glob:
+            _list = []
+            for obj in item_list:
+                if fnmatch.fnmatch(obj.title, title_glob):
+                    _list.append(obj)
+            item_list = OPItemList(_list)
         return item_list
 
     # TODO: Item creation is hard to test in an automated way since it results in changed
@@ -635,7 +649,8 @@ class OP(_OPCommandInterface, PyOPAboutMixin):
                           password: Union[str, OPPasswordRecipe] = None,
                           url: Optional[str] = None,
                           url_label: str = "Website",
-                          vault=None):  # pragma: no coverage
+                          tags: List[str] = [],
+                          vault: str = None):  # pragma: no coverage
         """
         Create a new login item in the authenticated 1Password account
 
@@ -653,6 +668,12 @@ class OP(_OPCommandInterface, PyOPAboutMixin):
             If an OPPasswordRecipe object is provided, it will ensure a well-formed password recipe string is provided to '--generate-password='
         url: str, optional
             If provided, set to the primary URL of the login item
+        url_label: str, optional
+            If provided and a URL is provided, this bcomes the primary URL's label
+        tags: List[str], optional
+            A list of tags to apply to the item when creating
+        vault: str, optionsl
+            The vault in which to create the new item
 
         Raises
         ------
@@ -681,8 +702,10 @@ class OP(_OPCommandInterface, PyOPAboutMixin):
         url_obj = None
         if url:
             url_obj = OPLoginItemNewPrimaryURL(url, url_label)
+
         new_item = OPLoginItemTemplate(
-            title, username, password=password, url=url_obj)
+            title, username, password=password, url=url_obj, tags=tags)
+
         login_item = self.item_create(
             new_item, password_recipe=password_recipe, vault=vault)
         return login_item
@@ -744,6 +767,92 @@ class OP(_OPCommandInterface, PyOPAboutMixin):
         self._item_delete(item_id, vault=vault, archive=archive)
 
         return item_id
+
+    def item_delete_multiple(self,
+                             vault,
+                             categories=[],
+                             include_archive=False,
+                             tags=[],
+                             archive=False,
+                             title_glob=None,
+                             batch_size=25):
+        """
+        Delete multiple items at once from a specific vault. This may take place across
+        one or more 'op item delete' passes. A maximum "batch size" number of items to
+        delete in each pass may optionally be specified (defaulting to 25)
+
+        Parameters
+        ----------
+        vault: str
+            The name or ID of a vault to delete from. This parameter is mandatory to mitigate
+            the risk of deleting many items from the wrong vault
+        include_archive: bool, optional
+            Whether to include archived items for deleting
+            by default False
+        archive: bool
+            Whether to archive or permanently delete the items
+            by default False
+        tags: List[str], optional
+            A list of tags to restrict batch deletion to
+        title_glob: bool, optional
+            a shell-style glob pattern to match against item titles for deleting
+            by default None
+        batch_size: int, optional
+            Maximum number of items to delete in each pass
+            by default 25
+            NOTE: The default batch size is subject to change without notice
+
+        Note:
+            If a non-unique item identifier is provided (e.g., item name/title), and there
+            is more than one item that matches, OPItemDeleteException will be raised. Check the
+            error message in OPItemDeleteException.err_output for details
+
+        Raises
+        ------
+        OPItemDeleteMultipleException
+            - If the 'item list' operation fails
+            - If any of the 'item delete' operations fail
+
+        Returns
+        -------
+        item_id: str
+            Unique identifier of the item deleted
+
+        """
+        # track deleted items as we delete them so we can return
+        # that list to the caller
+        deleted_items = OPItemList([])
+
+        try:
+            item_list = self.item_list(categories=categories,
+                                       include_archive=include_archive,
+                                       tags=tags,
+                                       title_glob=title_glob,
+                                       vault=vault)
+        except OPItemListException as e:
+            raise OPItemDeleteMultipleException.from_opexception(
+                e, deleted_items)
+
+        batches: List[OPItemList] = []
+        start = 0
+        end = len(item_list)
+
+        for i in range(start, end, batch_size):
+            # split item list into chunks >= batch_size each
+            x = i
+            chunk = item_list[x:x+batch_size]
+            batches.append(OPItemList(chunk))
+
+        for batch in batches:
+            batch_json = batch.serialize()
+            try:
+                self._item_delete_multiple(batch_json, vault, archive=archive)
+            except OPCmdFailedException as ope:  # pragma: no coverage
+                raise OPItemDeleteMultipleException.from_opexception(
+                    ope, deleted_items)
+            deleted_items.extend(batch)
+
+        return deleted_items
 
     def signed_in_accounts(self, decode="utf-8") -> OPAccountList:
         account_list_json = super()._signed_in_accounts(self.op_path, decode=decode)